Create hook for pre-check-in post and use for new pre check in flow
