+++ src/applications/check-in/components/ActionItemDisplay.jsx
@@ -1,36 +1,32 @@
 import React, { useMemo } from 'react';
 import { useSelector } from 'react-redux';
-import { useTranslation } from 'react-i18next';
 import PropTypes from 'prop-types';
 // eslint-disable-next-line import/no-unresolved
 import { recordEvent } from '@department-of-veterans-affairs/platform-monitoring/exports';
 import { createAnalyticsSlug } from '../utils/analytics';
 import { useFormRouting } from '../hooks/useFormRouting';
+
 import { makeSelectApp, makeSelectVeteranData } from '../selectors';
 import { APP_NAMES } from '../utils/appConstants';
 import {
-  hasPhoneAppointments,
   preCheckinAlreadyCompleted,
   getAppointmentId,
 } from '../utils/appointment';
 
 import WhatToDoNext from './WhatToDoNext';
+import PreCheckInSuccessAlert from './PreCheckInSuccessAlert';
 
 const ActionItemDisplay = props => {
   const { router } = props;
   const selectApp = useMemo(makeSelectApp, []);
   const { app } = useSelector(selectApp);
-  const { goToNextPage, jumpToPage } = useFormRouting(router);
+  const { goToNextPage, jumpToPage, pages } = useFormRouting(router);
   const selectVeteranData = useMemo(makeSelectVeteranData, []);
   const { appointments } = useSelector(selectVeteranData);
-  const { t } = useTranslation();
 
   const displaySuccessAlert =
-    app === APP_NAMES.PRE_CHECK_IN && preCheckinAlreadyCompleted(appointments);
-
-  const successMessage = hasPhoneAppointments(appointments)
-    ? t('your-provider-will-call-you-at-your-appointment-time')
-    : t('you-can-check-in-with-your-smartphone');
+    app === APP_NAMES.PRE_CHECK_IN &&
+    (preCheckinAlreadyCompleted(appointments) || pages.length < 5);
 
   const action = e => {
     e.preventDefault();
@@ -48,19 +44,7 @@ const ActionItemDisplay = props => {
   return (
     <>
       {displaySuccessAlert ? (
-        <section data-testid="pre-check-in-success-alert">
-          <va-alert
-            close-btn-aria-label="Close notification"
-            closeable
-            status="success"
-            visible
-          >
-            <h2 slot="headline">{t('your-information-is-up-to-date')}</h2>
-            <p data-testid="success-message" className="vads-u-margin-y--0">
-              {successMessage}
-            </p>
-          </va-alert>
-        </section>
+        <PreCheckInSuccessAlert />
       ) : (
         <section data-testid="what-to-do-next">
           <WhatToDoNext
