+++ src/applications/check-in/hooks/useFormRouting.jsx
@@ -64,6 +64,15 @@ const useFormRouting = (router = {}) => {
     [getCurrentPageFromRouter, pages],
   );
 
+  const getNextPageFromRouter = useCallback(
+    () => {
+      const currentPage = getCurrentPageFromRouter();
+      const positionInForm = pages.indexOf(currentPage);
+      return pages[positionInForm + 1];
+    },
+    [getCurrentPageFromRouter, pages],
+  );
+
   const goToNextPage = useCallback(
     () => {
       const here = getCurrentPageFromRouter();
@@ -81,6 +90,7 @@ const useFormRouting = (router = {}) => {
   return {
     getCurrentPageFromRouter,
     getPreviousPageFromRouter,
+    getNextPageFromRouter,
     goToErrorPage,
     jumpToPage,
     goToPreviousPage,
