+++ src/applications/check-in/locales/en/translation.json
@@ -279,5 +279,6 @@
   "phone-appointments": "Phone appointments",
   "check-in-now-for-your-date-time-appointment": "Check in now for appointment on {{ date, day }}, {{ date, monthDay }} at {{ date, time }}",
   "details-for-appointment-on-date-at-time": "Details for {{ type }} appointment on {{ dateTime, date }} at {{ dateTime, time }}",
-  "appointment-on-date-at-time": "Appointment on {{dateTime, date}} at {{ dateTime, time}}"
+  "appointment-on-date-at-time": "Appointment on {{dateTime, date}} at {{ dateTime, time}}",
+  "setting-pre-check-in-completed": "Setting Pre-Check-In completed"
 }
